Move flags package under a `launchdarkly` subdir
