+++ x/launchdarkly/flags/user_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/cultureamp/ca-go/request"
-	"github.com/cultureamp/ca-go/x/flags"
+	"github.com/cultureamp/ca-go/x/launchdarkly/flags"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/launchdarkly/go-sdk-common.v2/lduser"
